requirements.txt
google-adk==1.19.0
python-dotenv>=1.0.0
requests>=2.32.4
google-generativeai==0.8.5

google-cloud-bigquery
pandas
db-dtypes

google-auth
google-auth-oauthlib

pydantic
python-dateutil
PyYAML

anyio
fastapi
uvicorn
starlette
tenacity
typing-extensions
tzlocal
watchdog
websockets
